Fix TM2 unnecessary conversions (lint)

+++ traffic_monitor/experimental/traffic_monitor/enum/enum.go
@@ -201,7 +201,7 @@ func ToNumeric(v interface{}) (float64, bool) {
 	case float32:
 		return float64(i), true
 	case float64:
-		return float64(i), true
+		return i, true
 	case int:
 		return float64(i), true
 	case uint:
